Relocate geyser.pack -> geyser.pack.bedrock
