+++ core/src/main/java/org/geysermc/geyser/event/type/SessionLoadResourcePacksEventImpl.java
@@ -40,9 +40,9 @@ import org.geysermc.geyser.api.pack.UrlPackCodec;
 import org.geysermc.geyser.api.pack.exception.ResourcePackException;
 import org.geysermc.geyser.api.pack.option.PriorityOption;
 import org.geysermc.geyser.api.pack.option.ResourcePackOption;
-import org.geysermc.geyser.pack.GeyserResourcePack;
-import org.geysermc.geyser.pack.ResourcePackHolder;
-import org.geysermc.geyser.pack.option.OptionHolder;
+import org.geysermc.geyser.pack.bedrock.GeyserResourcePack;
+import org.geysermc.geyser.pack.bedrock.ResourcePackHolder;
+import org.geysermc.geyser.pack.bedrock.option.OptionHolder;
 import org.geysermc.geyser.registry.Registries;
 import org.geysermc.geyser.session.GeyserSession;
 import org.geysermc.geyser.util.GeyserIntegratedPackUtil;
