+++ core/src/main/java/org/geysermc/geyser/pack/bedrock/option/GeyserUrlFallbackOption.java
@@ -23,7 +23,7 @@
  * @link https://github.com/GeyserMC/Geyser
  */
 
-package org.geysermc.geyser.pack.option;
+package org.geysermc.geyser.pack.bedrock.option;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.geysermc.geyser.api.pack.ResourcePack;
