+++ core/src/main/java/org/geysermc/geyser/pack/bedrock/path/GeyserPathPackCodec.java
@@ -23,7 +23,7 @@
  * @link https://github.com/GeyserMC/Geyser
  */
 
-package org.geysermc.geyser.pack.path;
+package org.geysermc.geyser.pack.bedrock.path;
 
 import lombok.RequiredArgsConstructor;
 import org.checkerframework.checker.nullness.qual.NonNull;
