+++ core/src/main/java/org/geysermc/geyser/registry/loader/ResourcePackLoader.java
@@ -36,12 +36,12 @@ import org.geysermc.geyser.api.pack.ResourcePack;
 import org.geysermc.geyser.api.pack.ResourcePackManifest;
 import org.geysermc.geyser.api.pack.UrlPackCodec;
 import org.geysermc.geyser.event.type.GeyserDefineResourcePacksEventImpl;
-import org.geysermc.geyser.pack.GeyserResourcePack;
-import org.geysermc.geyser.pack.GeyserResourcePackManifest;
-import org.geysermc.geyser.pack.ResourcePackHolder;
-import org.geysermc.geyser.pack.SkullResourcePackManager;
-import org.geysermc.geyser.pack.path.GeyserPathPackCodec;
-import org.geysermc.geyser.pack.url.GeyserUrlPackCodec;
+import org.geysermc.geyser.pack.bedrock.GeyserResourcePack;
+import org.geysermc.geyser.pack.bedrock.GeyserResourcePackManifest;
+import org.geysermc.geyser.pack.bedrock.ResourcePackHolder;
+import org.geysermc.geyser.pack.bedrock.SkullResourcePackManager;
+import org.geysermc.geyser.pack.bedrock.path.GeyserPathPackCodec;
+import org.geysermc.geyser.pack.bedrock.url.GeyserUrlPackCodec;
 import org.geysermc.geyser.registry.Registries;
 import org.geysermc.geyser.session.GeyserSession;
 import org.geysermc.geyser.text.GeyserLocale;
